// newcode/libs/ui/common/src/lib/domain-controls/model/select-key-event.enum.ts
/*
 * $Id$
 * Copyright(c) RIB Software GmbH
 */

export enum SelectkeyCodes {
	BACKSPACE = 'Backspace',

	TAB = 'Tab',

	ENTER = 'Enter',

	ESCAPE = 'Escape',

	LEFT = 'ArrowLeft',

	UP = 'ArrowUp',

	RIGHT = 'ArrowRight',

	DOWN = 'ArrowDown',

	DELETE = 'Delete',

	SPACE = 'Space',

	PAGE_UP = 'PageUp',

	PAGE_DOWN = 'PageDown',
}